Add RemoteRepoCache.resolve_refs() method with tests
This method is for batch-resolving of refs into SHA1s using the remote
repo cache. It takes a list of (repo, ref) tuples and returns
dictionary that maps these tuples to dictionaries that look as follow
in successful cases

    {
        'repo': <original repo>,
        'ref': <original ref>,
        'repo-url': <expanded full repo URL>,
        'sha1': <resolved SHA1>,
        'tree': <tree SHA1 of the commit>,
    }

and like this in negative cases:

    {
        'repo': <original repo>,
        'ref': <original ref>,
        'repo-url': <expanded full repo URL>,
        'error': <error string>,
    }

The mapping may seem redundant but it might be useful when looking up
the results. If tuples are passed in they will be available at the
caller's end already.