Replace lambda with conditional
A visit function can optionally be given to traverse_morphs
to be called on each morph traverse_morph visits.

The current lambda definition has the wrong number of args.

Using a lambda as the default arg means that the number of
args the lambda takes has to be updated whenever the visit
function requires more args. Getting rid of the lambda
rids us of this extra maintenance.