BuildCommand: Validate multiple root morphologies
When you attempt to build a stratum or chunk, the ArtifactResolver can
return multiple root artifacts, since the root source produces multiple
artifacts.

Rather than having the BuildCommand complain that there's mutiple root
artifacts, it now validates all the produced artifacts too, since that
will validate the kinds of artifacts produced, and give a more useful
error message, that you're trying to build a Stratum or Chunk directly.

If all the produced artifacts validate, then an Exception is raised to
signal that it got multiple artifacts, when it only expected one.