CacheKeyComputer: Include split rules in computation
Chunk artifacts need the [(artifact_name, [regular_expression])] so that
if the default split rules change, or the blending rule changes, then
an extra version field doesn't need to be added to the cache key computer.

This is for future plans to allow the split rules to be configurable
and allow us to more easily change them.

System and Stratum artifact computations don't need this, since those
splitting rules are already expressed in the dependencies information.