Load morphologies from the definitions repo
Rather than using the repo given by a chunk or stratum in a stratum or
system, always look for morphologies in the definitions repository. It
is assumed that the repository of the first triplet given is the
definitions repository.

However, source creation must still use the repo and ref from the
chunk or stratum itself, so make sure that this is still available to use.