Build Scheduling Improvements
We replace the request queue with a pool of jobs.
When an initiator makes a build request for an artifact,
a Job is created for that artifact, if a job for the artifact
has already been created then we just add the initiator to the
list of initiators that are interested in this artifact. From
this point onwards the initiator will receive all build messages
and build output for that artifact.