Add handling for new messages to BuildController
There are two new messages:

WorkerBuildStepAlreadyStarted tells the initiator that the artifact
they want to build is already being built, e.g.

'eglibc-misc is already building on 172.17.1.37:3434'

WorkerBuildWaiting tells the initiator that the artifact they want
to build can't be built yet because there aren't any workers free, e.g.

'Ready to build eglibc-misc: waiting for a worker to become available'

In addition the contents of the message has changed for several events,
event messages that need to be sent to several initiators have a list
of ids instead of a single id.