Add command to validate the local artifact cache
This tests the checksums of artifacts and tests any unpacked chunks.

The unpacked chunks are tested by turning them back into tarfiles and
calculating a checksum, which we can rely on being identical to the
chunk artifact in the artifact cache if there is no corruption. That is
pretty cool!

This code doesn't work because the zlib.adler32() code can't be used as
a rolling checksum. I'm going to switch to using hashlib.md5() which
should still be fast enough. Alternately we could go for sha256 right
away, so that when we /do/ have bit-for-bit reproducibility from source,
we can say that everything is totally secure.