Allow the user to specify deployments in a cluster
Instead of taking the name of a cluster morphology and zero or more
parameters for overriding the cluster morphology, morph deploy should
take the name of a cluster morphology and the names of zero or more
system deployments that are defined in the cluster morphology. If no
deployment names are given then all deployments are deployed.

The parameters can still be passed in the same format as before using
the `--param="SYSTEM.KEY=VALUE"` option.