Don't validate when loading all morphologies
Currently, all morphologies in the definitions repo are validated when
we create a build branch, and when we load all of the morphologies for
traversal to change refs when using morph edit.

This commit stops the morphologies from being validated on load, and
instead validates only the morphologies that have been affected by
whatever was happening.