Raise exception if a ref is not valid in the remote repo cache
Previously if a ref didn't resolve in the remote cache, we'd clone the repo into the local
cache and check again. However, 99% of the time the clone comes from the remote cache so
we should be able to assume that if the ref doesn't exist there, it doesn't exist at all.