Don't assume morphologies end in .morph
We occasionally need to load every morphology in the repository, e.g.
morph edit needs to change every morphology that references that chunk.

We used to do this by loading everything that ended in `.morph`, but
if we're moving to allowing morphologies to have arbitrary filenames,
there's no suitable metric other than "this file will successfully
validate as a morphology".

So, the morphologyfinder module is no longer appropriate, and the
responsibility for working out what is a valid morphology has been
shifted to the SystemBranchDirectory class, which is used to having to
load all the morphologies.