Use exact filenames to refer to morphology files
Rather than repeatedly stripping and appending an optional .morph extension
morphology names, instead always use the file path of the morphology
relative to the definitions repository.

This is an inversion of the previous logic, which would strip the .morph
extension and use the "name" internally.

The exception to this rule of always using the filename, is that `morph
edit CHUNK` uses the name of the morphology as-defined in the stratum.

This is based off Adam Coldrick's inital patch, but this version will
allow the old style of providing the "name" by converting it into a path
if it does not have either a / or a . in it.

An unfortunate consequence of this change is that the show-dependencies
command's output changed, so the test needed updating.