Don't check if a file exists before trying to read it
We used to check whether a file existed before trying to read it. We
used to be able to get away with only looking at the top-level
directory, which made using ls-files before trying to cat-file it
better.

Unfortunately, we need to look at the files in subdirectories now, so
this no longer works.

We could make it include files in subdirectories, but for repositories
with many files, you would end up reading a file listing longer than
the morphology, so even in the slow case of needing to read the entire
morphology file, it would be faster to attempt to read the file first.

So now we beg forgiveness rather than asking permission.