yarns: Adapt to put morphologies in subdirs
We want to move our morphologies in our definitions repository into
subdirectories, so they're more organised.

We'd prefer to only refer to morphologies by file path, rather than a
name that loosely corresponds to the file path, but we need to support
that for backwards compatibility until we can move all of our
morphologies into the definitions repository.

However, since we want to eventually remove this, and we want to ensure
that file paths work, we change the yarn tests to use file paths.