Don't attempt to load symlinks as morphologies
This was noticed because our definitions.git had a dangling symlink,
so it failed to construct the temporary build branch.

We shouldn't process symlinks as morphologies either, since either we
make symlinked morphologies a first-class behaviour, and validate that
the link points inside the repository, which is a lot of work for
something we don't and probably won't need; or we can just assume that
we deal with the morphology this is linked to correctly anyway.