Next steps
I realise now that requiring the development and runtime dependencies of
everything just to construct a Gem is actually pretty wrong. Most of
them are the things you need to run the *tests*, which is a totally
different problem than just getting them to work.

It'd be great to use this info in the future to split the strata
appropriately (into -devel, -test and -runtime), but for now we can
probably just hardcode the build dependency info.