yarns: Don't use filepaths for chunk morphologies
The following commits will make morph attempt to load a chunk morphology
from the definitions repository if its morph field is a filepath rather
than a name. In preparation for this, this commit stops the yarn tests
from using a filepath in this field when the morphology is not in the
definitions repository.