Adapt traverse_morphs to be able to load chunk morphologies from definitions
This commit adapts the traverse_morphs code to handle keeping track of
both the morphology location and the source location for chunks by
loading morphologies from a given "morphology_repo" and "morphology_ref"
rather than the source repo and ref. The repo and ref of the source is
still passed and used as the reference for the chunk and given to the
Source object in the source pool.