Use the whole chunk morphology to compute cache keys
Change the cache key computer to use all (useful) fields in the chunk
morphology to compute the cache key for a chunk. Previously, changing a
chunk morphology meant the chunk ref would need to change, but now this
is not the case, so the cache key needs to change when the morphology
does.