Add scripts/import-rubygem to generate morphologies for RubyGems
This generates a stratum and a set of chunk morphologies for integrating
a given RubyGem project into Baserock. It uses the 'bundler' Gem to
collate the list dependencies. This means that it works for any Gem that
has a "Gemfile", and will also honour the "Gemfile.lock" file created by
Bundler if the file exists.

It requires the new 'rubygem' build mode.