Add optional 'gem-url' field in chunk morphologies
This could just as easily be a generic 'extra-build-data' field, which
could be used to download arbitrary files into the build chroot. Making
it easier to include random data from the internet in builds is not
really a step forward for a tool that tries to encourage reproducability
from source, however.