Add 'rubygem' build mode
This build mode honours the new 'gem-url' field. It downloads the
Gem from the given URL

Including code from random servers on the internet in your builds might
mean that you can't reproduce what you build from source later on,
because the file or even the whole server might be gone. Morph does
nothing to prevent this, either for Git source code or Ruby Gems.

The Trove server provides a way of mirroring Git repos on local
infrastructure, so organisations can be certain they won't disappear
later on and so that they are quicker to access. Trove does not yet
provide a similar service for mirroring the required RubyGems.