morph3: Add a get_commands method to Morphology
morph2.Morphology has a `get_commands` method which is used to return
the [configure|build|install]-commands of the chunk, either by getting
them from the chunk morphology or from the build system of the chunk.

This commit implements a similar method in morph3.Morphology and adds
tests for it, so that when we remove morph2 the API changes to
morphologies are minimal.