Bake 2: add rubygem.import script
This generates a chunk morph for ONE gem defined in a given git
repository.

It has one known issue: where gems are nested in a source repo (as is
the case in rails) they will not be found. The `bundle install` command
does find them, so clearly something is up with the way I am hacking the
'bundler' library around.

Also, there is lots of inheritance and method override hacking :(