Always use the current branch of edited chunks' repos
If you checkout a different branch, or even no branch at all with a
dangling head (such as bisecting) then morph will use that, rather than
what "ref" is set to in the morphology.

Push and temporary build branch creation avoidance still assume that the
ref in the edited chunk repository matches the ref in the morphology, so
if you checkout a different branch and everything is committed, it will
assume there's no local changes and build what the ref points to.