Replace sleeping non-blocking loop with asyncore
This is ~10 lines shorter than the non-blocking loop, though improving
the whitespace formatting would probably cut that down to being about
the same.

However asyncore uses a select/poll internally, so doesn't need to sleep
to avoid busy-waiting since it can block, which makes it more responsive
to output than the read loop.