Handle closing the ends of the log pipe we don't care about
We only need the read end in the parent, and we only need the write end
in the child, so we can close them pre-emptively.

Also, we need to close the write end in the parent so that when the
subprocess terminates we get EOF on the read end of the pipe.