import: Allow chunk morph generation to fail
Instead of exiting immediately, collect the errors up and report them
at the end. The tool should do as much as possible on each execution.

For example, this helps in the case where a user wants to get a rough
idea of how much is involved to integrate a given package.