import: Move the processing loop into its own class
This allows keeping the state as members instead of having to pass
12+ parameters to functions.

Also in this change: you can now specify 'extra_args' for a tool
which are passed at the start of the commandline. Some tools (e.g.
Omnibus) require more than just the source repo and name to generate
a chunk morph or lorry file.