stagingarea: Mount proc and ccache inside the namespace
This works towards allowing multiple concurrent builds in the same
system, which has the same problem as deployments.

This is the easy bit, since linux-user-chroot has support for bind
mounts and /proc mounts.

We also need to get rid of the /dev/shm mount to be able to build in
parallel though.