Update graphing plugin
At the moment the graphing plugin isn't too useful because it graphs every
artifact in the entire system, the resulting graph is too large to be useful.

This patch replace the graph-build-depends command with two new commands:
graph-sources and graph-artifacts. graph-sources constructs a graph of a
system's sources while graph-artifacts constructs a graph of a stratum's
artifacts (excluding artifacts that come from other strata i.e. strata
that the stratum we're graphing depends on)

Both commands omit edges between chunks that come from different strata
because this is implied by one stratum depending on another, however
both commands have a --show-all-edges option that can be used to override
this behaviour.