Run system-integrations in a namespace
All temporary mounts should be done inside a namespace, so that they
don't interfere with other namespaces.

system-integrations may need capabilities that regular builds don't
have, so they're chrooted, rather than linux-user-chrooted, which means
it's more complicated to do namespaces.

In the absence of a better command for it, we can do this with an
in-lined shell script.

This also stops us using the run-parts inside the system, and executes
the integrations directly.