Annotate dependencies to include runtime deps.
If you build-depend on a stratum, then you also build-depend on all that
stratum's runtime dependencies.

This means that the runtime dependencies will be made available at
build-time, so if you include a python module that only runtime depended
on a library, not build-depended, then its runtime dependency still gets
included.

Similar logic is used at system build time to ensure that runtime
dependencies of all the strata you include are also included.

This means that you can only list the strata that provide the
functionality you want for your system, rather than having to also list
the strata that it depends on.