Add initial implementation of an LRU cache for resolved tree refs
This uses the PyLRU module, from:
<https://pypi.python.org/pypi/pylru/1.0.6>.

Right now it's included in the source tree, but it'd be better to
include it as a separate chunk.

Python 3.2 and newer provide a built-in LRU cache, but this is
specifically for in-memory use. See <http://bugs.python.org/issue17528>.

Git commits are immutable, so caching information about their contents
is fairly easy and trouble-free. There's no danger of the cache becoming
stale.

<INSERT PERFORMANCE MEASUREMENTS HERE:
- use morph list-artifacts with
A) no repos cached locally, with and without the tree cache, x2
B) all repos cached locally, with and without the tree cache, x2
>