distbuild: Handle errors from socket
We found a distbuild controller stuck in a busy loop, with the logs
full of the same error message repeated:

    ... _flush(): Exception 'IOError: [Errno 32] Broken pipe' from sock.write()

We suspect this came about because the initiator disconnected without
sending an EOF. The initiator was in a VM on a laptop so it seems
possible that the host OS turned off the wireless adaptor without giving
the VM a chance to close its connections gracefully.

The busy loop is because nothing in the SocketBuffer class handles the
SocketError events queued by the _flush() method. Unhandled events are
ignored. So the SocketBuffer stays in 'w' state without ever shifting
any data and never returns. Adding transitions to handle the SocketError
event will fix the problem.

If a socket error happens now in the same scenario, it will be handled
as if the initiator disconnected.