Implement partial distbuilds
In addition to partial builds we also want to be able to do partial
distbuilds, and distbuild uses a different codepath.

This commit updates the distbuild code to know what to do if a partial
build is requested. It only builds up to the latest chunk/stratum that
was requested, and displays where to find the artifacts for each of
the chunks/strata requested upon completion of the build.

The usage is the same as for local builds.

Change-Id: I0537f74e2e65c7aefe5e71795f17999e2415fce5