Add plugin to get details of given chunk name.
Usefully, this provides the resolved repo URL, which may be
aliased in the actual defintions.  This makes it easier to
use a workflow that avoids `morph edit`.  Resolved repo URLs
can be git cloned.

It also provides other info about the chunk.  In the case
of multiple strata containing the same chunk name, the tool
lists them all, saying which stratum they come from.

Example output:

    # morph get-chunk-details libpng
    In stratum graphics-common:
      Chunk: libpng
       Repo: git://git.baserock.org/delta/libpng
        Ref: 259fb7761d747655c607efcec7a12ff1f3c24561

Change-Id: I6b540e38d9521f3f473b51c9031e508ddcb458ee