distbuild: Fix issues in build cancellation
A cancel during the 'graphing' or 'annotating' stages would be ignored
as the BuildController was listening for the InitiatorDisconnect message
from the wrong event source. In 'building' state the actual build would
be stopped, but the BuildController instance would stick around due to
sending the message class instead of an instance of the message.

Change-Id: I222a8aa39bf7fffab4d89e12997ffd18cd1b54fc