Add local cliapp module
Add a copy of the specific version of the cliapp module required by morph, a
known version of which can't otherwise be guaranteed to be available on
general Linux systems, or via PyPi. Including the module directly allows for
continued development in line with morph, until a time when it may become
desirable to move away from dependence on this module.

With a local copy of cliapp added to morph, it has been shown that morph is
capable of building Baserock images without depending on an existing Baserock
system.

Change-Id: Iab1f7037c7afff054a2404c0552d9b5e4d2d141f