CacheKeyComputer: add chunk-compatibility-version
Since we build chunks using the '-e' flag in `sh`, we might have
different artifacts as result of building a chunk. It's better
we change the cache-key of these so that we ensure that everything
rebuilds, and we don't find the case where something that doesn't
build, is already cached.

Change-Id: Ic0fe8f9cf36b7787ebb08bd8200e0f1ea4cb01ef